refactor: extract skill-check evaluation out of NarrativeController
Pulls the ~160-line skill-requirement block out of generateNextSegment into a
free function, evaluateDecisionSkillChecks. It's pure compute plus two injected
callbacks (toast, onSkillCheckPerformed) with no React state, so a free function
fits better than a hook. Behavior's unchanged: the controller resolves
character/world and calls the helper, with hasCriticalFailure staying inline
since it also needs decisionWeight.

First of two extractions chipping at the NarrativeController god-file. Adds a
focused unit test for the helper.
